move logic of VDI_Resolve() to VRT as VRT_DirectorResolve()
The use case is eager resolution of directors in vmods, for example to
determine if a director would resolve to a particular backend.

The functionality is identical to VDI_Resolve(), except that no state
in the busyobj is modified.

VDI_Resolve() remains for error handling for calls from VDI_GetHdr()
and VDI_Http1Pipe()